+++ src/flask_wtf/recaptcha/widgets.py
@@ -3,19 +3,16 @@ from urllib.parse import urlencode
 from flask import current_app
 from markupsafe import escape
 from markupsafe import Markup
+from wtforms.widgets import html_params
 
 RECAPTCHA_SCRIPT_DEFAULT = "https://www.google.com/recaptcha/api.js"
 RECAPTCHA_DIV_CLASS_DEFAULT = "g-recaptcha"
-RECAPTCHA_TEMPLATE = """
-<script src='{script}' async defer{nonce_attr}></script>
-<div class="{div_class}" {snippet}></div>
-"""
 
 __all__ = ["RecaptchaWidget"]
 
 
 class RecaptchaWidget:
-    def recaptcha_html(self, public_key, nonce=None):
+    def recaptcha_html(self, public_key, nonce=None, **kwargs):
         html = current_app.config.get("RECAPTCHA_HTML")
         if html:
             return Markup(html)
@@ -28,19 +25,22 @@ class RecaptchaWidget:
         if callable(nonce):
             nonce = nonce()
         nonce_attr = f' nonce="{escape(nonce)}"' if nonce else ""
-        attrs = current_app.config.get("RECAPTCHA_DATA_ATTRS", {})
-        attrs["sitekey"] = public_key
-        snippet = " ".join(f'data-{k}="{attrs[k]}"' for k in attrs)  # noqa: B028, B907
-        div_class = current_app.config.get("RECAPTCHA_DIV_CLASS")
-        if not div_class:
-            div_class = RECAPTCHA_DIV_CLASS_DEFAULT
+
+        kwargs.setdefault(
+            "class",
+            current_app.config.get("RECAPTCHA_DIV_CLASS")
+            or RECAPTCHA_DIV_CLASS_DEFAULT,
+        )
+
+        data_attrs = dict(current_app.config.get("RECAPTCHA_DATA_ATTRS", {}))
+        data_attrs["sitekey"] = public_key
+        for k, v in data_attrs.items():
+            kwargs.setdefault(f"data-{k}", v)
+
+        attributes = html_params(**kwargs)
         return Markup(
-            RECAPTCHA_TEMPLATE.format(
-                script=script,
-                nonce_attr=nonce_attr,
-                div_class=div_class,
-                snippet=snippet,
-            )
+            f"\n<script src='{script}' async defer{nonce_attr}></script>\n"
+            f"<div {attributes}></div>\n"
         )
 
     def __call__(self, field, error=None, **kwargs):
@@ -51,4 +51,5 @@ class RecaptchaWidget:
         except KeyError:
             raise RuntimeError("RECAPTCHA_PUBLIC_KEY config not set") from None
 
-        return self.recaptcha_html(public_key, nonce=field.nonce)
+        kwargs.setdefault("id", field.id)
+        return self.recaptcha_html(public_key, nonce=field.nonce, **kwargs)